+++ clang/unittests/StaticAnalyzer/CallDescriptionTest.cpp
@@ -9,6 +9,7 @@
 #include "Reusables.h"
 
 #include "clang/AST/ExprCXX.h"
+#include "clang/StaticAnalyzer/Core/PathSensitive/CallDescription.h"
 #include "clang/StaticAnalyzer/Core/PathSensitive/CallEvent.h"
 #include "clang/Tooling/Tooling.h"
 #include "gtest/gtest.h"
