+++ llvm/utils/gn/secondary/clang/lib/StaticAnalyzer/Core/BUILD.gn
@@ -20,6 +20,7 @@ static_library("Core") {
     "BlockCounter.cpp",
     "BugReporter.cpp",
     "BugReporterVisitors.cpp",
+    "CallDescription.cpp",
     "CallEvent.cpp",
     "Checker.cpp",
     "CheckerContext.cpp",
